forum page refactor
